requirements.txt
anyio==4.11.0
APScheduler==3.11.1
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-telegram-bot==22.5
sniffio==1.3.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
Werkzeug==3.1.3
